Aggiunto campo notifiche risposte per ogni utente salvato Aggiunta gestione eccezioni

+++ src/main/java/eu/Rationence/pat/model/User.java
@@ -55,7 +55,7 @@ public class User implements Serializable {
     private String description;
 
     @Column(name="i_Cost", nullable = false)
-    private double cost;
+    private int cost;
 
     @Column(name="x_Time", nullable = false, length=5)
     private String time;
